增加商品功能 --  01
详情： 初始化商品代码
原因：
解决方案：

+++ src/main/java/win/ots/hello/v1/service/IUserService.java
@@ -20,11 +20,11 @@ public interface IUserService extends IService<User> {
 
     User getByUserName(String userName);
 
-    Set<String> getRolesByUserId(Integer id);
+    Set<String> getRolesByUserId(Long id);
 
     List<User> getPageByQuery();
 
     UserInfoVo register(UserCreateVo createVo);
 
-    UserInfoVo getUserInfo(Integer userId);
+    UserInfoVo getUserInfo(Long userId);
 }
