+++ src/main/java/win/ots/hello/web/controller/UserController.java
@@ -1,18 +1,12 @@
 package win.ots.hello.web.controller;
 
 import org.apache.shiro.authz.annotation.RequiresRoles;
-import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
-import win.ots.hello.constant.ErrorMessage;
-import win.ots.hello.constant.RoleEnum;
-import win.ots.hello.constant.SexEnum;
 import win.ots.hello.constant.ShiroConstant;
-import win.ots.hello.core.exception.ServiceException;
 import win.ots.hello.core.result.Result;
 import win.ots.hello.core.result.ResultGenerator;
 import win.ots.hello.v1.entity.User;
-import win.ots.hello.v1.entity.UserRoleRelation;
 import win.ots.hello.v1.service.IUserRoleRelationService;
 import win.ots.hello.v1.service.IUserService;
 import win.ots.hello.web.query.UserQuery;
@@ -37,7 +31,7 @@ public class UserController {
 
     @RequiresRoles({ShiroConstant.ROLE_ADMIN})
     @RequestMapping(value = "/{user_id}", method = RequestMethod.GET)
-    public Result<User> get(@PathVariable(value = "user_id") int userId) {
+    public Result<User> get(@PathVariable(value = "user_id") Long userId) {
 
         User user = userService.getById(userId);
         return ResultGenerator.genSuccessResult(user);
@@ -51,7 +45,7 @@ public class UserController {
     }
 
     @RequestMapping(value = "/info/{user_id}", method = RequestMethod.GET)
-    public Result<UserInfoVo> getUserInfo(@PathVariable(value = "user_id") int userId) {
+    public Result<UserInfoVo> getUserInfo(@PathVariable(value = "user_id") Long userId) {
 
         UserInfoVo userInfo = userService.getUserInfo(userId);
         return ResultGenerator.genSuccessResult(userInfo);
